requirements.txt
flask
flask-cors
flask-pymongo
python-dotenv
pyjwt
bcrypt
sendgrid
requests
supabase
celery pymongo
pymongo[srv]
dnspython